requirements.txt
asgiref==3.8.1
Django==4.2
django-allauth==0.62.1
django-environ==0.11.2
django-livereload-server==0.5.1
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
sqlparse==0.5.0
tornado==6.4
